collectd-prv: escape trailing backslash
If line is chunked and the chunk ends in a backslash, the end quote will
be escaped and collectd will reject the notification with an error:

-1 Malformed option.

A trailing backslash may occur if there are escaped characters in the
stream:

INPUT:abc\n
CHUNKS:
@1:1:2@abc\
@1:2:2@n
OUTPUT:
message="@1:1:2@abc\"
message="@1:2:2@n"

What to do ... The backslash could be:

* removed (but the next message now begins with a mysterious 'n')

* escaped and the parser reconstructing the input must recognize the
  special case

* all backslashes escaped

* strip out all escaped characters from the input in collectd-prv

* strip out all escaped characters from the input in a pipeline (e.g.,
  using sed)

This change does option 2 and 5. The trailing backslash is escaped; if
the behaviour is not wanted, strip out the special characters before
sending to collect-prv.