Drop lines beginning with NUL
From OpenBSD's `fgets(3)`:

```
It is erroneous to assume that fgets() never returns an empty
string when successful.  If a line starts with the NUL character,
fgets will store the NUL and continue reading until it encounters a
newline or end-of-file.  This will result in an empty string being
returned.  The following bit of code illustrates a case where the
programmer assumes the string cannot be zero length.

      char buf[1024];

      if (fgets(buf, sizeof(buf), fp) != NULL) {
              /* WRONG */
              if (buf[strlen(buf) - 1] == '\n')
                      buf[strlen(buf) - 1] = '\0';
      }

If strlen() returns 0, the index into the buffer becomes -1.
```